refactor: sync directory structure
