+++ lib/features/sync/ui/widgets/matrix/status_indicator.dart
