+++ lib/features/sync/ui/widgets/matrix/verification_emojis_row.dart
