+++ lib/features/sync/ui/widgets/matrix/verification_modal.dart
@@ -1,61 +1,71 @@
 import 'dart:async';
 
-import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:lotti/features/sync/matrix.dart';
+import 'package:lotti/features/sync/ui/widgets/matrix/verification_emojis_row.dart';
 import 'package:lotti/l10n/app_localizations_context.dart';
 import 'package:lotti/providers/service_providers.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/buttons/lotti_primary_button.dart';
-import 'package:lotti/widgets/sync/matrix/verification_emojis_row.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
-import 'package:matrix/encryption.dart';
+import 'package:matrix/matrix.dart';
 
-class IncomingVerificationModal extends ConsumerStatefulWidget {
-  const IncomingVerificationModal(
-    this.keyVerification, {
+class VerificationModal extends ConsumerStatefulWidget {
+  const VerificationModal(
+    this.deviceKeys, {
     super.key,
   });
 
-  final KeyVerification keyVerification;
+  final DeviceKeys deviceKeys;
 
   @override
-  ConsumerState<IncomingVerificationModal> createState() =>
-      _IncomingVerificationModalState();
+  ConsumerState<VerificationModal> createState() => _VerificationModalState();
 }
 
-class _IncomingVerificationModalState
-    extends ConsumerState<IncomingVerificationModal> {
+class _VerificationModalState extends ConsumerState<VerificationModal> {
   MatrixService get _matrixService => ref.read(matrixServiceProvider);
+  KeyVerificationRunner? _runner;
+
+  @override
+  void dispose() {
+    _runner?.stopTimer();
+    super.dispose();
+  }
+
+  void startVerification() => _matrixService.verifyDevice(widget.deviceKeys);
+
+  @override
+  void initState() {
+    startVerification();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
     final pop = Navigator.of(context).pop;
 
-    void closeModal() {
-      Navigator.of(context).pop();
-    }
-
-    final unverifiedDevices = _matrixService.getUnverifiedDevices();
-    final requestingDevice = unverifiedDevices.firstWhereOrNull(
-      (deviceKeys) => deviceKeys.deviceId == widget.keyVerification.deviceId,
-    );
-
-    final displayName = requestingDevice?.deviceDisplayName ??
-        widget.keyVerification.deviceId ??
-        'device name not found';
-
     return StreamBuilder<KeyVerificationRunner>(
-      stream: _matrixService.incomingKeyVerificationRunnerStream,
+      stream: _matrixService.keyVerificationStream,
       builder: (context, snapshot) {
         final runner = snapshot.data;
+        _runner = runner;
         final lastStep = runner?.lastStep;
         final emojis = runner?.emojis;
+        final isLastStepKey = lastStep == 'm.key.verification.key';
         final isLastStepDone = lastStep == 'm.key.verification.done';
+        final isLastStepCancel = lastStep == 'm.key.verification.cancel';
 
         final isDone =
             isLastStepDone || (runner?.keyVerification.isDone ?? false);
+
+        if (isLastStepCancel) {
+          Timer(const Duration(seconds: 30), pop);
+        }
+
+        if (isLastStepDone) {
+          Timer(const Duration(seconds: 30), pop);
+        }
 
         return SingleChildScrollView(
           child: Padding(
@@ -68,22 +78,49 @@ class _IncomingVerificationModalState
                   children: [
                     Flexible(
                       child: Text(
-                        displayName,
+                        widget.deviceKeys.deviceDisplayName ??
+                            widget.deviceKeys.deviceId ??
+                            '',
                         style: context.textTheme.titleLarge,
                         softWrap: true,
                       ),
                     ),
                   ],
                 ),
+                const SizedBox(height: 10),
+                Opacity(
+                  opacity: 0.5,
+                  child: Text(
+                    widget.deviceKeys.userId,
+                    style: context.textTheme.bodySmall,
+                  ),
+                ),
                 const SizedBox(height: 20),
-                if (!isDone && emojis == null)
+                if (runner == null)
                   LottiPrimaryButton(
-                    onPressed: runner?.acceptVerification,
-                    label: 'Verify Session',
+                    key: const Key('matrix_start_verify'),
+                    onPressed: startVerification,
+                    label:
+                        context.messages.settingsMatrixStartVerificationLabel,
+                  ),
+                if (lastStep?.isEmpty ?? false)
+                  Text(
+                    context.messages.settingsMatrixContinueVerificationLabel,
+                  ),
+                if (isLastStepCancel)
+                  Text(
+                    context.messages.settingsMatrixVerificationCancelledLabel,
+                  ),
+                if (isLastStepKey && emojis == null)
+                  LottiPrimaryButton(
+                    key: const Key('matrix_accept_verify'),
+                    onPressed: runner?.acceptEmojiVerification,
+                    label:
+                        context.messages.settingsMatrixAcceptVerificationLabel,
                   ),
                 if (!isDone && emojis != null) ...[
                   Text(
-                    context.messages.settingsMatrixVerifyIncomingConfirm,
+                    context.messages.settingsMatrixVerifyConfirm,
                     style: context.textTheme.bodyLarge,
                     textAlign: TextAlign.center,
                   ),
@@ -98,7 +135,8 @@ class _IncomingVerificationModalState
                         child: LottiPrimaryButton(
                           key: const Key('matrix_cancel_verification'),
                           onPressed: () async {
-                            closeModal();
+                            await runner?.cancelVerification();
+                            pop();
                           },
                           label: context
                               .messages.settingsMatrixCancelVerificationLabel,
@@ -119,8 +157,8 @@ class _IncomingVerificationModalState
                 if (isDone) ...[
                   Text(
                     context.messages.settingsMatrixVerificationSuccessLabel(
-                      '',
-                      runner?.keyVerification.deviceId ?? '',
+                      widget.deviceKeys.deviceDisplayName ?? '',
+                      widget.deviceKeys.deviceId ?? '',
                     ),
                   ),
                   const SizedBox(height: 20),
@@ -150,48 +188,5 @@ class _IncomingVerificationModalState
         );
       },
     );
-  }
-}
-
-class IncomingVerificationWrapper extends ConsumerStatefulWidget {
-  const IncomingVerificationWrapper({super.key});
-
-  @override
-  ConsumerState<IncomingVerificationWrapper> createState() =>
-      _IncomingVerificationWrapperState();
-}
-
-class _IncomingVerificationWrapperState
-    extends ConsumerState<IncomingVerificationWrapper> {
-  StreamSubscription<KeyVerification>? _subscription;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _subscription = ref
-        .read(matrixServiceProvider)
-        .getIncomingKeyVerificationStream()
-        .listen((keyVerification) {
-      if (mounted) {
-        showModalBottomSheet<void>(
-          context: context,
-          builder: (context) {
-            return IncomingVerificationModal(keyVerification);
-          },
-        );
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    _subscription?.cancel();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox.shrink();
   }
 }
