+++ lib/main.dart
@@ -10,6 +10,7 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/database/logging_db.dart';
 import 'package:lotti/database/maintenance.dart';
 import 'package:lotti/database/settings_db.dart';
+import 'package:lotti/database/sync_db.dart';
 import 'package:lotti/features/ai/repository/ai_config_repository.dart'
     hide aiConfigRepositoryProvider;
 import 'package:lotti/features/sync/matrix/matrix_service.dart';
@@ -100,6 +101,7 @@ Future<void> main() async {
           maintenanceProvider.overrideWithValue(getIt<Maintenance>()),
           journalDbProvider.overrideWithValue(getIt<JournalDb>()),
           loggingDbProvider.overrideWithValue(getIt<LoggingDb>()),
+          syncDatabaseProvider.overrideWithValue(getIt<SyncDatabase>()),
           loggingServiceProvider.overrideWithValue(getIt<LoggingService>()),
           outboxServiceProvider.overrideWithValue(getIt<OutboxService>()),
           aiConfigRepositoryProvider
