+++ test/features/sync/ui/unverified_devices_page_test.dart
@@ -3,12 +3,13 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/features/sync/matrix/matrix_service.dart';
 import 'package:lotti/features/sync/state/matrix_unverified_provider.dart';
 import 'package:lotti/features/sync/ui/unverified_devices_page.dart';
+import 'package:lotti/features/sync/ui/widgets/matrix/device_card.dart';
+import 'package:lotti/features/sync/ui/widgets/matrix/status_indicator.dart';
 import 'package:lotti/providers/service_providers.dart';
-import 'package:lotti/widgets/sync/matrix/device_card.dart';
-import 'package:lotti/widgets/sync/matrix/status_indicator.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:matrix/matrix.dart';
 import 'package:mocktail/mocktail.dart';
+
 import '../../../widget_test_utils.dart';
 
 class _FakeMatrixUnverifiedController extends MatrixUnverifiedController {
