+++ test/features/sync/ui/widgets/matrix/device_card_test.dart
@@ -1,12 +1,12 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lotti/features/sync/matrix/matrix_service.dart';
+import 'package:lotti/features/sync/ui/widgets/matrix/device_card.dart';
 import 'package:lotti/providers/service_providers.dart';
-import 'package:lotti/widgets/sync/matrix/device_card.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:matrix/matrix.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../../widget_test_utils.dart';
+import '../../../../../widget_test_utils.dart';
 
 class MockMatrixService extends Mock implements MatrixService {}
