+++ test/features/sync/ui/widgets/outbox_monitor_test.dart
@@ -9,8 +9,8 @@ import 'package:lotti/features/user_activity/state/user_activity_service.dart';
 import 'package:lotti/get_it.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../mocks/sync_config_test_mocks.dart';
-import '../../widget_test_utils.dart';
+import '../../../../mocks/sync_config_test_mocks.dart';
+import '../../../../widget_test_utils.dart';
 
 void main() {
   var syncDatabaseMock = MockSyncDatabase();
